pyproject.toml
[project]
name = "rag-es"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "elasticsearch>=9.2.1",
    "huggingface-hub[hf-xet]>=0.36.0",
    "langchain>=1.2.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langsmith>=0.5.2",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
